Add a relevant element that was forgotten.
On the file show_cupons.html.erb the value of :coupon was put inside.
The area for show the price in show_cupon, now show the cupon of descoutn too.
The way to defind the type of value for coupon in the forms was changed of url_field to text_field.
With the new element the design was adapted to for show_cupon.